SendSS: ! fixed wrong printf format. (can cause a crash on x86_64)
git-svn-id: http://svn.miranda-ng.org/main/trunk@9451 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/SendScreenshotPlus/src/UMainForm.cpp
@@ -238,7 +238,7 @@ void TfrmMain::wmInitdialog(WPARAM wParam, LPARAM lParam) {
 	ComboBox_SetCurSel (hCtrl,0);
 	if(m_MonitorCount >1) {
 		TCHAR	tszTemp[120];
-		for (size_t i = 0; i < m_MonitorCount; ++i) {
+		for (int i = 0; i < m_MonitorCount; ++i) {
 			mir_sntprintf(tszTemp, SIZEOF(tszTemp),_T("%i. %s%s"),
 				i+1,
 				TranslateT("Monitor"),
